Rename game states to scenes
