+++ src/game/soldier.c
@@ -22,7 +22,7 @@ float soldier_scale = 1.0;
 int is_swimming = 0;
 
 // Variable externa que necesitamos
-extern int game_state;
+extern int game_scene;
 extern float[MaxBullets] bullet_x;
 extern float[MaxBullets] bullet_y;
 extern float[MaxBullets] bullet_angle;
@@ -142,7 +142,7 @@ void update_soldier()
                 mark_event_as_happened(GameOver);
             }
             soldier_scale = SoldierDrownScale;
-            game_state = StateGameOver;
+            game_scene = SceneGameOver;
         }
         else
         {
@@ -438,7 +438,7 @@ void soldier_take_damage()
 
             mark_event_as_happened(GameOver);
         }
-        game_state = StateGameOver;
+        game_scene = SceneGameOver;
     }
     else
     {
