+++ src/scenes/gameover.c
@@ -2,14 +2,14 @@
 #include "../utils/definitions.h"
 #include "../game/heli.h"
 
-extern int game_state;
+extern int game_scene;
 
 void update_gameover() {
     // Volver al menú si se presiona START
     if(gamepad_button_start() == 1) {
         reset_heli();  // Reiniciar el estado del avión
         reset_soldier();
-        game_state = StateGame;
+        game_scene = SceneGame;
         
         if (!has_event_happened(GameRestart))
         {
@@ -22,6 +22,19 @@ void update_gameover() {
     }
 }
 
+int[2][15] TextsGameOver =
+{
+    "  GAME OVER",
+    "FIN DEL JUEGO"
+};
+
+int[2][30] TextsContinue =
+{
+    " PRESS START TO CONTINUE",
+    "PULSA START PARA CONTINUAR"
+};
+
+
 void render_gameover() {
     // Limpiar la pantalla con negro opaco
     clear_screen(BackgroundColor);
@@ -31,12 +44,12 @@ void render_gameover() {
     set_multiply_color(RedColor);
     
     // Mostrar mensaje de game over
-    print_at(ScreenCenterX - 45, ScreenCenterY - 20, "GAME OVER");
+    print_at(ScreenCenterX - 65, ScreenCenterY - 20, TextsGameOver[game_language]);
     
     // Hacer parpadear el texto usando el contador de frames
     if((get_frame_counter() / BlinkRate) % 2 == 0) {
         set_multiply_color(TextColor);
-        print_at(ScreenCenterX - 115, ScreenCenterY + 20, "PRESS START TO CONTINUE");
+        print_at(ScreenCenterX - 125, ScreenCenterY + 20, TextsContinue[game_language]);
     }
     set_multiply_color(color_white);
 }
